Backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
python-dotenv
sqlalchemy
psycopg2-binary
chromadb
openai
PyMuPDF
langchain-text-splitters
python-multipart
